python_requirements.txt
pandas==1.1.3
numpy==1.21.1
pyarrow==5.0.0
sklearn==0.24.1
pytorch=1.7.0
